fix build errors

+++ endpoint/endpoint.go
@@ -96,7 +96,7 @@ func MakeKeysAndStats(rows []map[string]bigquery.Value) ([]*datastore.Key, []Sta
 // FetchEndpointStats executes simpleQuery, and returns a slice of rows containing
 // endpoint signatures and request counts.
 // TODO - move the body (excluding simpleQuery) into go/bqext
-func FetchEndpointStats(dsExt *bqext.Dataset, threshold int64) ([]map[string]bigquery.Value, error) {
+func FetchEndpointStats(dsExt *bqext.Dataset, threshold int) ([]map[string]bigquery.Value, error) {
 	qString := strings.Replace(simpleQuery, "${THRESHOLD}", fmt.Sprint(threshold), 1)
 	qString = strings.Replace(qString, "${DATE}", fmt.Sprint(threshold), 1)
